Made to follow stricter C standards.

+++ src/get_block.c
@@ -1,13 +1,15 @@
 #include <stdio.h>
 #include <string.h>
 
+#include <get_block.h>
+
 int get_block_cbraces(char *str, int *start, int *end) {
     char *b1 = strchr(str, '{');
     if(b1 == NULL) {
         fprintf(stderr, "get_block_cbraces: Could not locate first curly brace!\n");
         return 1;
     }
-    *start = (b1 - str);
+    *start = (int)(b1 - str);
     int bn = 1;
     b1++;
     while(*b1 != '\0') {
@@ -15,7 +17,7 @@ int get_block_cbraces(char *str, int *start, int *end) {
         else if(*b1 == '}') bn--;
         b1++;
         if(bn == 0) {
-            *end = (b1 - str);
+            *end = (int)(b1 - str);
             return 0;
         }
     }
